Fix Certificate storage lookup on FreeBSD
Previously on FreeBSD the path /usr/local/share/certs was found and
caused the main lookup loop to exit. Since this directory only
contains one certificate bundle and not single certificates, the loop
exited and no certificates were loaded. By moving this line, the root
certs are recognized correctly on FreeBSD.

Confirmed to be working on FreeBSD 12.1-RELEASE-p8.